Updates knowing that the audio and video timestamps are 64-bit values
Previously, I had suspected this, but I have now been able to confirm
it on actual data where the timestamp "wrapped" in the 32-bit sense.

+++ riff/types.go
@@ -125,7 +125,7 @@ type Chunk struct {
 	ID         string
 	Data       []byte
 	IsKeyframe bool   // Custom; this is true if we want to index this chunk.
-	Timestamp  uint32 // Custom; this helps us order our chunks.
+	Timestamp  uint64 // Custom; this helps us order our chunks.
 }
 
 // AVIChunkIndex describes a chunk index.
